feat: change prices to double

+++ Audentia-Core/src/main/java/fr/audentia/core/domain/model/shop/ShopItem.java
@@ -3,9 +3,9 @@ package fr.audentia.core.domain.model.shop;
 public class ShopItem {
 
     public final String material;
-    public final int price;
+    public final double price;
 
-    public ShopItem(String material, int price) {
+    public ShopItem(String material, double price) {
         this.material = material;
         this.price = price;
     }
